+++ model/user.go
@@ -9,11 +9,11 @@ type (
 	}
 
 	User struct {
-		Role    Role    `gorm:"column:role" json:"role"`       // 用户角色
-		Balance float64 `gorm:"column:balance" json:"balance"` // 余额
-		Phone   string  `gorm:"column:phone" json:"phone"`     // 电话号码
-		OpenID  string  `gorm:"column:openid" json:"openid"`   // 微信 OpenID
-		Credit  int     `gorm:"column:credit" json:"credit"`   // 信誉分
+		Role    Role   `gorm:"column:role" json:"role"`       // 用户角色
+		Balance int64  `gorm:"column:balance" json:"balance"` // 余额，单位：分
+		Phone   string `gorm:"column:phone" json:"phone"`     // 电话号码
+		OpenID  string `gorm:"column:openid" json:"openid"`   // 微信 OpenID
+		Credit  int    `gorm:"column:credit" json:"credit"`   // 信誉分
 	}
 
 	Role int
